Create new script file with changes
With this new file you can string together several operations; however,
if you want to start over after the square root or the percent,
you need to press clear. Now all the features work fine, they just need
the visual feedback and the rounding to limit the number of digits.